+++ examples/local_project/lib/docker_project_web/controllers/page_controller.ex
@@ -1,0 +1,7 @@
+defmodule LocalProjectWeb.PageController do
+  use LocalProjectWeb, :controller
+
+  def index(conn, _params) do
+    render(conn, "index.html")
+  end
+end