+++ examples/local_project/lib/docker_project_web/router.ex
@@ -1,0 +1,26 @@
+defmodule LocalProjectWeb.Router do
+  use LocalProjectWeb, :router
+
+  pipeline :browser do
+    plug :accepts, ["html"]
+    plug :fetch_session
+    plug :fetch_flash
+    plug :protect_from_forgery
+    plug :put_secure_browser_headers
+  end
+
+  pipeline :api do
+    plug :accepts, ["json"]
+  end
+
+  scope "/", LocalProjectWeb do
+    pipe_through :browser
+
+    get "/", PageController, :index
+  end
+
+  # Other scopes may use custom stacks.
+  # scope "/api", LocalProjectWeb do
+  #   pipe_through :api
+  # end
+end