+++ examples/local_project/lib/docker_project_web/views/layout_view.ex
@@ -1,0 +1,3 @@
+defmodule LocalProjectWeb.LayoutView do
+  use LocalProjectWeb, :view
+end