+++ examples/local_project/lib/docker_project_web/views/page_view.ex
@@ -1,0 +1,3 @@
+defmodule LocalProjectWeb.PageView do
+  use LocalProjectWeb, :view
+end